Some adjustments to the ruler code that might fix Bug #2363, @MrB, please test and adjust as needed.
git-svn-id: svn://scribus.net/trunk/Scribus@12347 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/hruler.h
@@ -101,7 +101,7 @@ signals:
 
 private:
 	double iter, iter2;
-	int cor;
+	double cor;
 	int whereToDraw;
 	bool drawMark;
 	PrefsManager *prefsManager;
